apps/backend/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.7
gunicorn==20.1.0
